dev-requirements.txt
pytest
pytest-cov
pre-commit
ruff
mypy
PyYAML
